QVAC-24422 ggml: integer TDT step control row that carries the emitted pair
The greedy transducer step control produced an F32 row holding only integers,
so every consumer paid for a conversion: the frame index needed a ggml_cast
before ggml_get_rows, and the emitted (token, duration index) pair had to be
copied out with two ggml_cpy nodes. On a per-dispatch-bound backend those three
extra nodes cost more than the step itself.

The row is now GGML_TYPE_I32 end to end, together with the loop counters and the
duration table, and it gained TOKEN and DUR slots that the kernel fills from its
own inputs. The frame slot is then already the index type ggml_get_rows wants
and the pair needs no copy.

The result is a view of a caller-supplied destination, the way ggml_cpy returns
a view of its own destination, so a chain of steps writes consecutive rows of
one buffer that the host reads back in a single transfer. The destination is
carried in src[4] so the graph reaches the buffer and the allocator, the
scheduler and the backends' range-based synchronisation all see the aliasing
they already handle for ggml_cpy.

The masked LSTM cell reads its update flag from that row, so its mask is I32 in
every backend as well.

CPU, CUDA, Metal and Vulkan kernels are updated together and stay line-for-line
identical; test-backend-ops covers the new layout, the written pair fields and a
destination view at a non-zero row offset.

(cherry picked from commit bef04319d3a1d91bdbbca6b4199bb6c5fe46e1aa)

+++ src/ggml-cuda/lstm-cell.cu
@@ -6,7 +6,7 @@
 // unmasked and hc_prev [2H, N] masked, where a zero mask entry copies the
 // column's previous pair through untouched.
 static __global__ void lstm_cell_f32_kernel(
-        const float * gates, const float * prev, const float * mask, float * dst,
+        const float * gates, const float * prev, const int * mask, float * dst,
         const int64_t H, const int64_t prev_row, const int64_t c_base,
         const int64_t mask_stride, const int64_t total) {
     const int64_t k = (int64_t) blockIdx.x * blockDim.x + threadIdx.x;
@@ -20,7 +20,7 @@ static __global__ void lstm_cell_f32_kernel(
     const float * p   = prev + n*prev_row;
     float       * out = dst  + n*GGML_LSTM_N_OUTS*H;
 
-    if (mask && mask[n*mask_stride] == 0.0f) {
+    if (mask && mask[n*mask_stride] == 0) {
         out[GGML_LSTM_OUT_H*H + j] = p[j];
         out[GGML_LSTM_OUT_C*H + j] = p[c_base + j];
         return;
@@ -62,6 +62,6 @@ void ggml_cuda_op_lstm_cell(ggml_backend_cuda_context & ctx, ggml_tensor * dst) 
     const int64_t num_blocks = (total + CUDA_LSTM_CELL_BLOCK_SIZE - 1) / CUDA_LSTM_CELL_BLOCK_SIZE;
     lstm_cell_f32_kernel<<<num_blocks, CUDA_LSTM_CELL_BLOCK_SIZE, 0, ctx.stream()>>>(
         (const float *) gates->data, (const float *) prev->data,
-        mask ? (const float *) mask->data : nullptr, (float *) dst->data,
+        mask ? (const int *) mask->data : nullptr, (float *) dst->data,
         H, prev->ne[0], c_base, (mask && ggml_nelements(mask) > 1) ? 1 : 0, total);
 }
